testing: keep the interest-cache disarm out of flui-foundation
The reviewer is right, and the contract is explicit:
`crates/flui-foundation/AGENTS.md` says "Emission only, never installation …
Subscriber construction lives in `flui-log`", and names the boundary as one
issue #568 deliberately restored. `impl Subscriber for InterestSentinel` plus
`Dispatch::new` is subscriber construction, in foundation. Feature-gating does
not change the owning layer, and neither does the fact that the *dependency*
half of the rationale was untouched: the module imported only `tracing`, and
foundation's manifest never gained `tracing-subscriber` or any sink.

So the primitive goes back to `flui-testing`, where it belongs — it is test
support — and is now public as `log_capture::disarm_interest_cache` for the
crates whose own capture helper is too specialised to replace. They reach it
through a dev-dependency cycle: `flui-testing` depends on them normally, and
cargo permits the reverse edge for dev-dependencies precisely so a lower crate
can use the test support built on it. Verified end to end, not assumed —
`cargo check --all-targets` and the full suite pass for all four.

`flui-foundation`'s own capture drops the call rather than acquiring one. Its
two capture tests emit at their own distinct source lines inside their own
closures, so no other test can reach those callsites first — the same reason
`flui-log` was left alone. Nothing is lost by leaving it out.

What the previous commit got right stands: the consumers keep their own
subscribers and assertions and add one call. What it got wrong was the home.
