+++ crates/flui-app/src/app/ui_realm.rs
@@ -10238,8 +10238,8 @@ mod tests {
             // Disarm `tracing`'s process-global callsite-interest cache first: it is
             // computed on whichever thread reaches a callsite FIRST, so without this a
             // sibling test can have it cached as `never` and silently empty this capture.
-            // See `flui_foundation::tracing_interest`.
-            flui_foundation::tracing_interest::disarm_interest_cache();
+            // See `flui_testing::log_capture`.
+            flui_testing::log_capture::disarm_interest_cache();
             tracing::subscriber::with_default(subscriber, || {
                 assert!(realm.render_frame_entered(&mut backend));
             });
