+++ crates/flui-foundation/src/diagnostics.rs
@@ -86,11 +86,6 @@ mod tests {
         }
 
         let capture = FieldCapture::default();
-        // Disarm `tracing`'s process-global callsite-interest cache first: it is
-        // computed on whichever thread reaches a callsite FIRST, so without this a
-        // sibling test can have it cached as `never` and silently empty this capture.
-        // See `flui_foundation::tracing_interest`.
-        crate::tracing_interest::disarm_interest_cache();
         tracing::subscriber::with_default(Registry::default().with(capture.clone()), emit);
 
         capture
