+++ crates/flui-view/Cargo.toml
@@ -53,11 +53,12 @@ dyn-clone = "1.0"
 tracing-subscriber = { workspace = true, optional = true }
 
 [dev-dependencies]
-# `testing` for `tracing_interest::disarm_interest_cache`, which this crate's
-# tracing-capturing tests call so a sibling test cannot poison their
-# callsites' cached interest. Dev-only: the feature never reaches a
-# production build.
-flui-foundation = { path = "../flui-foundation", version = "0.2.0", features = ["testing"] }
+# `log_capture::disarm_interest_cache`, which this crate's tracing-capturing
+# tests call so a sibling test cannot poison their callsites' cached interest.
+# A dev-only edge, and a dev-only CYCLE: flui-testing depends on this crate
+# normally, which cargo permits for dev-dependencies precisely so a lower crate
+# can use the test support built on it.
+flui-testing = { path = "../flui-testing", version = "0.2.0" }
 criterion = { workspace = true }
 # Enable flui-rendering's `testing` feature for unit tests so the
 # layout-builder tests can read committed geometry back off the pipeline
