+++ docs/runtime-contract.toml
@@ -233,7 +233,6 @@ declarations = [
     "pub mod listener_registry;",
     "pub mod notifier_generic;",
     "pub mod debug;",
-    "pub mod tracing_interest;",
     "pub use affinity::OwnerAffinity;",
     "pub use async_snapshot::{AsyncSnapshot, ConnectionState};",
     "pub use callbacks::{ FallibleCallback, Predicate, ValueChanged, ValueGetter, ValueSetter, ValueTransformer, VoidCallback, };",
@@ -282,7 +281,7 @@ declarations = [
     "pub mod replay;",
     "pub use a11y::{A11yNode, A11yQuery, A11yQueryError, A11yTree, NotTreeBound};",
     "pub use bootstrap::{BuildCapabilities, MountOptions, MountOwners, Mounted};",
-    "pub use log_capture::{CapturedLog, CapturedRecord, capture};",
+    "pub use log_capture::{CapturedLog, CapturedRecord, capture, disarm_interest_cache};",
     "pub use replay::{GestureRecorder, PointerPhase, PointerScript, ScriptedPointer};",
     "pub struct HeadlessBinding {",
 ]
